Moved get parameters to helper.js
